完善歌单详情界面,添加NeteaseLoadingDelegate抽象类,方便实现loading动画

+++ imooc_voice/src/main/java/com/imooc/imooc_voice/model/newapi/PlaylistDetailBean.java
@@ -696,7 +696,7 @@ public class PlaylistDetailBean {
             private int gender;
             private int city;
             private long birthday;
-            private int userId;
+            private long userId;
             private int userType;
             private String nickname;
             private String signature;
@@ -788,7 +788,7 @@ public class PlaylistDetailBean {
                 this.birthday = birthday;
             }
 
-            public int getUserId() {
+            public long getUserId() {
                 return userId;
             }
